requirements.txt
flask
flask_restful
requests
python-dotenv
pytest
requests-mock
